fix: harden NVENC encode path for reliability
- Fix NVENC session leak on cuCtxPopCurrent failure in nvCreateContext
- Fix coded buffer object leak on bitstreamData allocation failure
- Fix EOS flush ordering: flush encoder before freeing output buffer
- Fix integer overflow in rate control bitrate calculation (uint32 * uint32)
- Fix nvPutImage to respect src/dest offset parameters per VA-API spec
- Add missing bitrate extraction from HEVC sequence parameters
- Remove dead NVENCInputSurface struct and unused macros
- Remove unused drv parameter from nvRenderPictureEncode
- Normalize naming convention: hevc_enc_ -> hevcenc_ to match h264enc_

+++ src/h264_encode.c
@@ -81,7 +81,7 @@ void h264enc_handle_misc_params(NVENCContext *nvencCtx, NVBuffer *buffer)
         if (rc->bits_per_second > 0) {
             nvencCtx->maxBitrate = rc->bits_per_second;
             if (rc->target_percentage > 0) {
-                nvencCtx->bitrate = rc->bits_per_second * rc->target_percentage / 100;
+                nvencCtx->bitrate = (uint32_t)((uint64_t)rc->bits_per_second * rc->target_percentage / 100);
             } else {
                 nvencCtx->bitrate = rc->bits_per_second;
             }
